pyproject.toml
[project]
name = "protocols"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    # "aioquic>=1.2.0",
    "asyncio>=4.0.0",
    "fastapi>=0.116.1",
    "grpcio>=1.74.0",
    "grpcio-tools>=1.74.0",
    "httpx[http2]>=0.28.1",
    "hypercorn[h3]>=0.17.3",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "protobuf>=6.31.1",
    "uvicorn[standard]>=0.35.0",
]
